audio_evals/lib/cv3_speaker_sim/requirements.txt
addict
datasets==2.16.0
editdistance>=0.5.2
funasr
jiwer
kaldiio
librosa
numpy
onnxruntime-gpu
openai-whisper
pandas
pillow
pytest-runner
pytorch_wpe
PyYAML>=5.1.2
scipy
simplejson
sortedcontainers
torchcodec
tqdm
transformers
zhconv
zhon
